custom functions for addCourse and addBookmarks
